refactor: split app into modules
